Feat: useTitlePath hook
