+++ components/Element/Admin/AdminHeader.tsx
@@ -1,11 +1,17 @@
 import styled from 'styled-components';
 import { AdminHeaderProps } from '../../../@types/shared';
+import { Fonts } from '../../../styles/fonts';
 import { Sizes } from '../../../styles/sizes';
+import useTitle from '../../../hooks/useTitlePath';
 
 export default function AdminHeader({ onToggleSidebar }: AdminHeaderProps) {
+	const { title } = useTitle();
 	return (
 		<S.AdminHeaderLayout>
-			<button onClick={onToggleSidebar}>닫기버튼</button>
+			<div>
+				<button onClick={onToggleSidebar}>닫기버튼</button>
+				<h1>{title}</h1>
+			</div>
 			<div>마크</div>
 		</S.AdminHeaderLayout>
 	);
@@ -20,9 +26,19 @@ namespace S {
 		align-items: center;
 		justify-content: space-between;
 
-		> button {
-			background-color: #e2536b;
-			padding: 1rem;
+		> div {
+			display: flex;
+			align-items: center;
+			gap: 1rem;
+
+			> h1 {
+				${Fonts.bold20}
+			}
+
+			> button {
+				background-color: #e2536b;
+				padding: 1rem;
+			}
 		}
 	`;
 }
